+++ pypeit/core/procimg.py
@@ -14,58 +14,58 @@ from pypeit import debugger
 # TODO: Add sigdev to the high-level parameter set so that it can be
 # changed by the user?
 # JFH I think this crappy code below is deprecated
-def find_bad_pixels(bias, numamplifiers, datasec, sigdev=10.0, trim=True):
-    """
-    Identify bad pixels in the datasection of the bias frame based on
-    their robust deviation from the median.
-
-    Args:
-        bias (:obj:`numpy.ndarray`):
-            Bias frame
-        numamplifiers (int):
-            Number of amplifiers
-        datasec (list):
-            List of slices, one per amplifier, that contain the data in
-            the raw frame.  The slices and be lists of slice ojects or
-            strings.  If they are strings, :func:`parse.sec2slice` is
-            used to convert them for use in the function.
-        sigdev (:obj:`float`, optional):
-            Number of robust standard deviations beyond which to flag
-            pixels as bad.
-        trim (:obj:`bool`, optional):
-            Flag to trim image to the data section.
-
-    Returns:
-        :obj:`numpy.ndarray`: An integer array with bad pixels set to 1.
-
-    Raises:
-        ValueError:
-            Raised if the number of data sections does not match the
-            number of amplifiers.
-    """
-    # Check input
-    if len(datasec) != numamplifiers:
-        raise ValueError('Number of amplifiers does not match provided data sections.')
-
-    # If the input image sections are strings, convert them
-    if isinstance(datasec[0], str):
-        _datasec = datasec.copy()
-        for i in range(numamplifiers):
-            _datasec[i] = parse.sec2slice(datasec[i], require_dim=2)
-    else:
-        _datasec = datasec
-
-    # Find the bad pixels and mask them
-    mask = np.zeros_like(bias, dtype=np.int8)
-    is_data = np.zeros_like(bias, dtype=bool)
-    for i in range(numamplifiers):
-        is_data[datasec[i]] = True
-        temp = np.abs(np.median(bias[datasec[i]])-bias[datasec[i]])
-        sigval = 1.4826*max(np.median(temp), 1)
-        mask[datasec[i]][temp > sigdev*sigval] = 1
-
-    msgs.info("Identified {0:d} bad pixels".format(int(np.sum(mask))))
-    return trim_frame(mask, np.invert(is_data)) if trim else mask
+#def find_bad_pixels(bias, numamplifiers, datasec, sigdev=10.0, trim=True):
+#    """
+#    Identify bad pixels in the datasection of the bias frame based on
+#    their robust deviation from the median.
+#
+#    Args:
+#        bias (:obj:`numpy.ndarray`):
+#            Bias frame
+#        numamplifiers (int):
+#            Number of amplifiers
+#        datasec (list):
+#            List of slices, one per amplifier, that contain the data in
+#            the raw frame.  The slices and be lists of slice ojects or
+#            strings.  If they are strings, :func:`parse.sec2slice` is
+#            used to convert them for use in the function.
+#        sigdev (:obj:`float`, optional):
+#            Number of robust standard deviations beyond which to flag
+#            pixels as bad.
+#        trim (:obj:`bool`, optional):
+#            Flag to trim image to the data section.
+#
+#    Returns:
+#        :obj:`numpy.ndarray`: An integer array with bad pixels set to 1.
+#
+#    Raises:
+#        ValueError:
+#            Raised if the number of data sections does not match the
+#            number of amplifiers.
+#    """
+#    # Check input
+#    if len(datasec) != numamplifiers:
+#        raise ValueError('Number of amplifiers does not match provided data sections.')
+#
+#    # If the input image sections are strings, convert them
+#    if isinstance(datasec[0], str):
+#        _datasec = datasec.copy()
+#        for i in range(numamplifiers):
+#            _datasec[i] = parse.sec2slice(datasec[i], require_dim=2)
+#    else:
+#        _datasec = datasec
+#
+#    # Find the bad pixels and mask them
+#    mask = np.zeros_like(bias, dtype=np.int8)
+#    is_data = np.zeros_like(bias, dtype=bool)
+#    for i in range(numamplifiers):
+#        is_data[datasec[i]] = True
+#        temp = np.abs(np.median(bias[datasec[i]])-bias[datasec[i]])
+#        sigval = 1.4826*max(np.median(temp), 1)
+#        mask[datasec[i]][temp > sigdev*sigval] = 1
+#
+#    msgs.info("Identified {0:d} bad pixels".format(int(np.sum(mask))))
+#    return trim_frame(mask, np.invert(is_data)) if trim else mask
 
 
 #def badpix(frame, numamplifiers, datasec, sigdev=10.0):
@@ -506,9 +506,11 @@ def rn_frame(datasec_img, gain, ronoise, numamplifiers=1):
                              mask=indx).filled(0.0)
 
 
-def subtract_overscan(rawframe, numamplifiers, datasec, oscansec, method='savgol', params=[5, 65]):
+def subtract_overscan(rawframe, numamplifiers, datasec, oscansec, method='savgol', params=[5,65]):
     """
     Subtract overscan
+
+    TODO: Describe the method.
 
     Args:
         frame (:obj:`numpy.ndarray`):
@@ -517,15 +519,14 @@ def subtract_overscan(rawframe, numamplifiers, datasec, oscansec, method='savgol
             Number of amplifiers for this detector.
         datasec (list):
             List of slices, one per amplifier, that contain the data in
-            the raw frame.  The slices and be lists of slice ojects or
-            strings.  If they are strings, :func:`parse.sec2slice` is
-            used to convert them for use in the function.
+            the raw frame.  The slices must be lists of slice ojects.
+            See :func:`parse.sec2slice` to convert a string section (as
+            read from a file header) into a list of slices.
         oscansec (list):
-            List of slices, one per amplifier, that contain the
-            overscane regions in the raw frame.  The slices and be lists
-            of slice ojects or strings.  If they are strings,
-            :func:`parse.sec2slice` is used to convert them for use in
-            the function.
+            List of slices, one per amplifier, that contain the overscan
+            regions in the raw frame.  The slices must be lists of slice
+            ojects.  See :func:`parse.sec2slice` to convert a string
+            section (as read from a file header) into a list of slices.
         method (:obj:`str`, optional):
             The method used to fit the overscan region.  Options are
             polynomial, savgol, median.
@@ -541,27 +542,18 @@ def subtract_overscan(rawframe, numamplifiers, datasec, oscansec, method='savgol
     """
     # Check input
     if len(datasec) != numamplifiers or len(oscansec) != numamplifiers:
-        raise ValueError('Number of amplifiers does not match provided image sections.')
+        msgs.error('Number of amplifiers does not match provided image sections.')
 
     # If the input image sections are strings, convert them
-    if isinstance(datasec[0], str):
-        _datasec = datasec.copy()
-        for i in range(numamplifiers):
-            _datasec[i] = parse.sec2slice(datasec[i], require_dim=2)
-    else:
-        _datasec = datasec
-            
-    if isinstance(oscansec[0], str):
-        _oscansec = oscansec.copy()
-        for i in range(numamplifiers):
-            _oscansec[i] = parse.sec2slice(oscansec[i], require_dim=2)
-    else:
-        _oscansec = oscansec
+    if not isinstance(datasec[0], slice):
+        msgs.error('Data section must be a list of slice objects.')
+    if not isinstance(oscansec[0], slice):
+        msgs.error('Overscan section must be a list of slice objects.')
     
     # Check that there are no overlapping data sections
     testframe = np.zeros_like(rawframe, dtype=int)
     for i in range(numamplifiers):
-        testframe[_datasec[i]] += 1
+        testframe[datasec[i]] += 1
     if np.any(testframe > 1):
         raise ValueError('Image has overlapping data sections!')
 
@@ -571,10 +563,10 @@ def subtract_overscan(rawframe, numamplifiers, datasec, oscansec, method='savgol
     # Perform the bias subtraction for each amplifier
     for i in range(numamplifiers):
         # Pull out the overscan data
-        overscan = rawframe[_oscansec[i]]
+        overscan = rawframe[oscansec[i]]
 
         # Shape along at least one axis must match
-        data_shape = rawframe[_datasec[i]].shape
+        data_shape = rawframe[datasec[i]].shape
         if not np.any([ dd == do for dd, do in zip(data_shape, overscan.shape)]):
             msgs.error('Overscan sections do not match amplifier sections for'
                        'amplifier {0}'.format(i+1))
@@ -591,13 +583,13 @@ def subtract_overscan(rawframe, numamplifiers, datasec, oscansec, method='savgol
             ossub = signal.savgol_filter(osfit, params[1], params[0])
         elif method.lower() == 'median':
             # Subtract scalar and continue
-            nobias[_datasec[i]] -= osfit
+            nobias[datasec[i]] -= osfit
             continue
         else:
             raise ValueError('Unrecognized overscan subtraction method: {0}'.format(method))
 
         # Subtract along the appropriate axis
-        nobias[_datasec[i]] -= (ossub[:,None] if compress_axis == 1 else ossub[None,:])
+        nobias[datasec[i]] -= (ossub[:,None] if compress_axis == 1 else ossub[None,:])
 
     return nobias